Loop

+++ C++BASIC QUESTION/AreaOfCircle.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    int radius,area;
+    float radius,area;
     cout<<"enter the radius of circle : ";
     cin>>radius;
     area=3.14*radius*radius;
